requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
